// frontend/src/lib/api/orval/api/generated/model/errorCorrectionExercise.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Langu-AI API
 * API for the Langu-AI German language learning platform
 * OpenAPI spec version: 1.0.0
 */
import type { ErrorCorrectionExerciseExerciseType } from "./errorCorrectionExerciseExerciseType";

export interface ErrorCorrectionExercise {
  readonly id: string;
  exercise_type: ErrorCorrectionExerciseExerciseType;
  readonly item_id: string;
  readonly item_text: string;
  readonly item_translation: string;
  sentence: string;
  error_start: number;
  error_end: number;
  correct_replacement: string;
  corrected_sentence: string;
  explanation?: string;
}